[BACKEND] Fix layouts in nvgpu RewriteTensorPointer.
This pass was creating invalid IR.  In particular, it would create an
rank-N tensor with a rank-M layout, where N != M.

It "happened to work" because a later pass (probably remove layout
conversions?) was fixing it up.

Example of invalid IR: https://github.com/openai/triton/pull/2622#issuecomment-1803139094

This was caught by an IR-verifier rule I'm working on that specifically
checks that the rank of a layout matches that of its tensor.

GPC: rewrite-tensor-pointer-layouts